+++ scubainit/scubainit.c
@@ -39,9 +39,13 @@
 #define SCUBAINIT_VERBOSE   "SCUBAINIT_VERBOSE"
 
 static bool m_verbose = false;
-static int m_uid = -1;
-static int m_gid = -1;
-static int m_umask = -1;
+
+#define NO_VALUE    -1
+#define HAS_VALUE(x)    ((x) != NO_VALUE)
+
+static int m_uid = NO_VALUE;
+static int m_gid = NO_VALUE;
+static int m_umask = NO_VALUE;
 
 static const char *m_user;
 static const char *m_group;
@@ -63,16 +67,16 @@ path_join(const char *p1, const char *p2)
 }
 
 
-/* Returns true if root should be used */
+/* Returns true if scubainit will be changing users */
 static bool
-use_root(void)
+should_change_user(void)
 {
-    if (m_uid == -1) {
-        assert(m_gid == -1);
+    if (HAS_VALUE(m_uid)) {
+        assert(HAS_VALUE(m_gid));
         return true;
     }
 
-    assert(m_gid >= 0);
+    assert(!HAS_VALUE(m_gid));
     return false;
 }
 
@@ -112,14 +116,22 @@ add_group(const char *path, const char *name, unsigned int gid)
         if (gr == NULL)
             break;
 
-        if (strcmp(gr->gr_name, name) == 0) {
-            errmsg("Group \"%s\" already exists in %s\n", name, path);
+        bool name_matches = (strcmp(gr->gr_name, name) == 0);
+        bool gid_matches = (gr->gr_gid == gid);
+
+        if (name_matches) {
+            if (gid_matches) {
+                /* Identical name+gid exists; surprising, but no problem */
+                result = 0;
+                goto out;
+            }
+            errmsg("Group \"%s\" already exists with different gid in %s\n",
+                    name, path);
             goto out;
         }
 
-        if (gr->gr_gid == gid) {
-            errmsg("GID %u already exists in %s\n", gid, path);
-            goto out;
+        if (gid_matches) {
+            errmsg("Warning: GID %u already exists in %s\n", gid, path);
         }
     }
 
@@ -184,14 +196,22 @@ add_user(const char *path, const char *name, unsigned int uid, unsigned int gid,
         if (pw == NULL)
             break;
 
-        if (strcmp(pw->pw_name, name) == 0) {
-            errmsg("User \"%s\" already exists in %s\n", name, path);
+        bool name_matches = (strcmp(pw->pw_name, name) == 0);
+        bool uid_matches = (pw->pw_uid == uid);
+
+        if (name_matches) {
+            if (uid_matches) {
+                /* Identical name+uid exists; surprising, but no problem */
+                result = 0;
+                goto out;
+            }
+            errmsg("User \"%s\" already exists with different uid in %s\n",
+                    name, path);
             goto out;
         }
 
-        if (pw->pw_uid == uid) {
-            errmsg("UID %u already exists in %s\n", uid, path);
-            goto out;
+        if (uid_matches) {
+            errmsg("Warning: UID %u already exists in %s\n", uid, path);
         }
     }
 
@@ -258,7 +278,8 @@ add_shadow(const char *path, const char *name)
             break;
 
         if (strcmp(sp->sp_namp, name) == 0) {
-            errmsg("User \"%s\" already exists in %s\n", name, path);
+            /* Already exists; we don't really care about its values */
+            result = 0;
             goto out;
         }
     }
@@ -294,6 +315,10 @@ out:
 static int
 change_user(const char *home)
 {
+    assert(HAS_VALUE(m_uid));
+    assert(HAS_VALUE(m_gid));
+    assert(m_user != NULL);
+
     /* TODO: Would we ever want to get this list from scuba, too? */
     if (setgroups(0, NULL) != 0) {
         errmsg("Failed to setgroups(): %m\n");
@@ -407,6 +432,33 @@ out:
     return ret;
 }
 
+
+static void
+handle_wait_status(const char *cmd, int wstatus)
+{
+    if (wstatus < 0) {
+        errmsg("Failed to execute %s: %m\n", cmd);
+        exit(99);
+    }
+    else if (WIFEXITED(wstatus)) {
+        int es = WEXITSTATUS(wstatus);
+        if (es != 0) {
+            errmsg("%s exited with status %d\n", cmd, es);
+            exit(99);
+        }
+        // Success
+    }
+    else if (WIFSIGNALED(wstatus)) {
+        int sig = WTERMSIG(wstatus);
+        errmsg("%s terminated by signal %d\n", cmd, sig);
+        exit(99);
+    }
+    else {
+        errmsg("%s exited for an unknown reason! (wstatus=0x%X)\n", cmd, wstatus);
+        exit(99);
+    }
+}
+
 static int
 call_hook(const char *hook_path)
 {
@@ -423,16 +475,7 @@ call_hook(const char *hook_path)
 
     verbose("About to execute %s\n", hook_path);
     rc = system(hook_path);
-
-    if (rc < 0) {
-        errmsg("Failed to execute %s: %m\n", hook_path);
-        exit(99);
-    }
-    if (rc > 0) {
-        errmsg("%s exited with status %d\n", hook_path, rc);
-        exit(rc);
-    }
-
+    handle_wait_status(hook_path, rc);
     return 0;
 }
 
@@ -603,7 +646,7 @@ main(int argc, char **argv)
     if (process_envvars() < 0)
         exit(99);
 
-    if (!use_root()) {
+    if (should_change_user()) {
         /* Create user home directory */
         home = path_join(USER_HOME, m_user);
         if (make_homedir(home, m_uid, m_gid) != 0)
@@ -623,7 +666,7 @@ main(int argc, char **argv)
     call_hook(m_root_hook);
 
     /* Handle the scuba user */
-    if (!use_root()) {
+    if (should_change_user()) {
         if (change_user(home) < 0)
             goto fail;
 
@@ -637,7 +680,7 @@ main(int argc, char **argv)
     }
 
     /* Call post-su hook, only if we switch users */
-    if (!use_root()) {
+    if (should_change_user()) {
         call_hook(m_user_hook);
     }