+++ CHANGELOG.md
@@ -4,7 +4,35 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Changed
+- Switched to using `argcomplete` to provide Bash command line completion (#162)
+
+### Fixed
+- scubainit ignores matching passwd/group/shadow file entries when creating the
+  user. This allows transparently running scuba as root. (#164)
+- Fixed bug where scubainit incorrectly displayed the exit status of a failed
+  hook script. (#165)
+
+
+## [2.6.0] - 2020-03-25
+### Added
+- Add ability to override the shell in which the scuba-generated
+  script is run, via command line option (`--shell`) or via
+  `.scuba.yml` (#159)
+- Add ability to specify in `.scuba.yml` that a particular alias
+  should run as root (#159)
+
+
+## [2.5.0] - 2020-03-05
+### Changed
 - Use username/groupname of invoking user inside container (#153)
+- Ignore already existing UID/GIDs (#139)
+- Allow top-level `image` to be omitted in `.scuba.yml` (#158)
+
+### Fixed
+- Fixed deprecation error with collections.Mapping (#156)
+
+### Removed
+- Drop support for Python 2 (#154)
 
 
 ## [2.4.2] - 2020-02-24
@@ -191,7 +219,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 First versioned release
 
 
-[Unreleased]: https://github.com/JonathonReinhart/scuba/compare/v2.4.2...HEAD
+[Unreleased]: https://github.com/JonathonReinhart/scuba/compare/v2.6.0...HEAD
+[2.6.0]: https://github.com/JonathonReinhart/scuba/compare/v2.5.0...v2.6.0
+[2.5.0]: https://github.com/JonathonReinhart/scuba/compare/v2.4.2...v2.5.0
 [2.4.2]: https://github.com/JonathonReinhart/scuba/compare/v2.4.1...v2.4.2
 [2.4.1]: https://github.com/JonathonReinhart/scuba/compare/v2.4.0...v2.4.1
 [2.4.0]: https://github.com/JonathonReinhart/scuba/compare/v2.3.0...v2.4.0